Add files via upload
Advanced OpenCV
1. translation, rotation, flipping of images to certain distance and angles and directions 
2. BGR is the default setting for an image in OpenCV. it can be converted to RGB, HSV, LAB, and grayscale. 
But to convert the LAB/grayscale image to HSV we have to convert it first into BGR
3. in Splitmerge, we split the BGR color channels of the image with cv.split code. the shape of the split image is grayscale with different pixel distributions. 
So in order to actually see the color channel we first create a blank image and then split the color channels. 
#regions where it is lighter there is more pixel value 
#regions where it is darker there is less pixel value
to merge all the color channels cv.merge is used.
4. In smoothing.py, we have used diff methods to blurr the image. blurring thee image is basically reducing the noise of the image. 
to blurr the image a karnal window is drawn on one section of the image. the karnal window has a size which consists of rows and column. 
In the method of blurring, pixel intensity is reduced by methods like averaging, gausian blur, median blur and Bilateral Blurring