Add files via upload
When doing the face recognition and face training, we use the OpenCV library for Haarcascade which has the XML file for frontal face recognition. 
for Face training, two npy files are made for features that are the face features of different images and the labels for the same face of the character. 
The Face_trained.yml file is used for face recognition after training the computer to recognize different faces.